Implement local file read option.
During develop and general iteration, it seems a shame to redownload the
same file N times.  This is both slow, and doesn't play nice with NCBI's
bandwidth.  Further, if ever a researcher really *does* want to use this
script, iterating a search over a local file is much, much faster.  Win
win.